[breaking change] Stop using k3s on top of tailscale because the setup is flaky
Manual intervention required: remove the nodes and rejoin them with the
new ip one by one. Alternatively, nuke the cluster and recreate it from
backup.

Every once in a while, the inter-node connection would become unresponsive, leading to downtime. This is due to tailscale reconfiguring its wireguard mesh.

Until a better solution is found, it is best to stop using the tailscale interface and use the LAN interface instead.

The tailscale setup was only needed when a multi-site connectivity was needed anyway.

ref: https://chatgpt.com/share/692b67ba-8c48-8008-b228-25e112d284f2